LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/anonymous-iframe/anonymous-iframe-popup.tentative.https.window-expected.txt


FAIL Cross-origin popup from normal/anonymous iframes. assert_equals: Opener from anonymous iframe should be blocked. expected null but got object "[object Window]"
FAIL Same-origin popup from normal/anonymous iframes. assert_equals: Opener from anonymous iframe should be blocked. expected null but got object "[object Window]"
